[fix] #34 - 코드리뷰 반영
