+++ POTI-iOS/POTI-iOS/Presentation/Recruit/Enum/Section.swift
@@ -1,0 +1,12 @@
+//
+//  Section.swift
+//  POTI-iOS
+//
+//  Created by 이서현 on 1/15/26.
+//
+
+enum Section: Int, CaseIterable {
+    case recruitInfo
+    case progress
+    case participantInfo
+}
