+++ POTI-iOS/POTI-iOS/Presentation/Recruit/Enum/UserRole.swift
@@ -1,0 +1,11 @@
+//
+//  UserRole.swift
+//  POTI-iOS
+//
+//  Created by 이서현 on 1/15/26.
+//
+
+enum UserRole {
+    case host
+    case participant
+}
